Restructuring directory
Changing directory pathnames for eclipse compatibility
