+++ library-management/src/lab4/library/Book.java
@@ -1,0 +1,12 @@
+package lab4.library;
+ 
+public class Book {
+
+	// GRASP Principle: Information Expert
+	// Manages its own state (available or borrowed).
+	private String title;
+	private boolean isAvailable; 
+	
+	// TODO: implement Book functionality 
+	
+}
