+++ library-management/src/lab4/library/Member.java
@@ -1,0 +1,13 @@
+package lab4.library;
+
+public class Member {
+
+	// GRASP Principle: Information Expert and Low Coupling
+	// Can borrow and return books.
+
+	private String name;
+	// private borrowedbBooks TODO: implement collection of borrowed books
+	
+	// TODO: implement functionality of Member class
+
+}
