--- library/LibrarianController.java
@@ -1,11 +1,0 @@
-package lab4.library;
-
-public class LibrarianController {
-
-	// GRASP Principle: Controller
-	// Handles user requests like borrowing and returning books.
-	// Delegates tasks to the appropriate objects.
-	
-	// TODO: implement functionality of Member class
-
-}
