--- library/Library.java
@@ -1,16 +1,0 @@
-package lab4.library;
-
-public class Library {
-	
-	//  GRASP Principles: Creator, Controller
-	//	Manage the catalog of books and members.
-	//	Track which books are borrowed and available
-	
-	
-	// Attributes:
-	// private catalog (list of Book) 
-	// private members (list of Members)
-
-	// TODO: implement functionality of Member class
-
-}
